Unify GPU detection behind hw_detect and extract accel.py
Consolidate duplicated GPU detection logic so hw_detect.py is the
single source of truth for hardware detection, and extract all
acceleration selection code (check_*, get_accel, CDI validation,
accel_image) from common.py into a focused accel.py module.

- hw_detect.py: extend DeviceInfo with index/uuid fields, add
  per-detector cached functions, use Vulkan API directly via ctypes
  instead of parsing vulkaninfo output, add ARM guard to AmdRocmDetector,
  add Intel iGPU device IDs
- accel.py: new module with GPU selection, env var setup, CDI validation,
  and container image selection (moved from common.py)
- common.py: shed ~250 lines of acceleration code, now general utilities
- Update 14 import sites across the codebase
- Split acceleration tests into test_accel.py
