+++ ramalama/common.py
@@ -1,34 +1,30 @@
 from __future__ import annotations
 
-import glob
 import hashlib
 import json
 import os
-import platform
 import random
 import re
 import shutil
 import string
 import subprocess
 import sys
-from collections.abc import Callable, Sequence
+from collections.abc import Sequence
 from dataclasses import dataclass
 from functools import lru_cache
 from pathlib import Path
-from typing import IO, TYPE_CHECKING, Any, Literal, Optional, Protocol, TypedDict, Union, cast, get_args
+from typing import IO, TYPE_CHECKING, Any, Optional, Union
 
 if TYPE_CHECKING:
-    from typing_extensions import TypeAlias
-import yaml
+    pass
 
-import ramalama.amdkfd as amdkfd
 from ramalama.logger import logger
 from ramalama.version import version
 
 if TYPE_CHECKING:
     from argparse import Namespace
 
-    from ramalama.arg_types import SUPPORTED_ENGINES, ContainerArgType
+    from ramalama.arg_types import SUPPORTED_ENGINES
     from ramalama.config import Config
     from ramalama.transports.base import Transport
 
@@ -41,7 +37,6 @@ MNT_CHAT_TEMPLATE_FILE = f"{MNT_DIR}/chat_template.file"
 RAG_DIR = "/rag"
 RAG_CONTENT = f"{MNT_DIR}/vector.db"
 
-MIN_VRAM_BYTES = 1073741824  # 1GiB
 
 SPLIT_MODEL_PATH_RE = r'(.*?)(?:/)?([^/]*)-00001-of-(\d{5})\.gguf'
 
@@ -306,340 +301,6 @@ def engine_version(engine: SUPPORTED_ENGINES | Path | str) -> SemVer:
     return SemVer.parse(run_cmd(cmd_args, encoding="utf-8").stdout.strip())
 
 
-class CDI_DEVICE(TypedDict):
-    name: str
-
-
-class CDI_RETURN_TYPE(TypedDict):
-    devices: list[CDI_DEVICE]
-
-
-def load_cdi_config(spec_dirs: list[str]) -> Optional[CDI_RETURN_TYPE]:
-    # Load and merge all CDI configuration files from the given directories.
-    # When multiple CDI configs exist (e.g. /var/run/cdi and /etc/cdi), all are
-    # merged so that device "all" and other devices are found regardless of
-    # which file they appear in (fixes #2485).
-    merged_devices: list[CDI_DEVICE] = []
-    seen_names: set[str] = set()
-
-    for spec_dir in spec_dirs:
-        if not os.path.isdir(spec_dir):
-            continue
-        for root, _, files in os.walk(spec_dir):
-            for file in sorted(files):
-                _, ext = os.path.splitext(file)
-                if ext not in (".yaml", ".yml", ".json"):
-                    continue
-                file_path = os.path.join(root, file)
-                config = None
-                try:
-                    with open(file_path, "r") as stream:
-                        if ext == ".json":
-                            config = json.load(stream)
-                        else:
-                            config = yaml.safe_load(stream)
-                except (OSError, yaml.YAMLError, json.JSONDecodeError, UnicodeDecodeError) as e:
-                    logger.warning(f"Failed to load/parse CDI file {file_path}: {e}")
-                    continue
-                if not config or not isinstance(config, dict):
-                    continue
-                for cdi_device in config.get("devices", []):
-                    if isinstance(cdi_device, dict) and (name := cdi_device.get("name")):
-                        if name not in seen_names:
-                            seen_names.add(name)
-                            merged_devices.append(cast(CDI_DEVICE, cdi_device))
-
-    if not merged_devices:
-        return None
-    return {"devices": merged_devices}
-
-
-def get_podman_machine_cdi_config() -> Optional[CDI_RETURN_TYPE]:
-    cdi_config = run_cmd(["podman", "machine", "ssh", "cat", "/etc/cdi/nvidia.yaml"], encoding="utf-8").stdout.strip()
-    if cdi_config:
-        return yaml.safe_load(cdi_config)
-    return None
-
-
-def find_in_cdi(devices: list[str]) -> tuple[list[str], list[str]]:
-    # Attempts to find a CDI configuration for each device in devices
-    # and returns a list of configured devices and a list of
-    # unconfigured devices.
-    if platform.system() == "Windows":
-        cdi = get_podman_machine_cdi_config()
-    else:
-        cdi = load_cdi_config(['/var/run/cdi', '/etc/cdi'])
-    try:
-        cdi_devices = cdi.get("devices", []) if cdi else []
-        cdi_device_names = [name for cdi_device in cdi_devices if (name := cdi_device.get("name"))]
-    except (AttributeError, KeyError, TypeError) as e:
-        # Malformed YAML or JSON. Treat everything as unconfigured but warn.
-        logger.warning(f"Unable to process CDI configuration: {e}")
-        return ([], devices)
-
-    configured = []
-    unconfigured = []
-    for device in devices:
-        if device in cdi_device_names:
-            configured.append(device)
-        # A device can be specified by a prefix of the uuid
-        elif device.startswith("GPU") and any(name.startswith(device) for name in cdi_device_names):
-            configured.append(device)
-        else:
-            perror(f"Device {device} does not have a CDI configuration")
-            unconfigured.append(device)
-
-    return configured, unconfigured
-
-
-def check_asahi() -> Optional[Literal["asahi"]]:
-    if os.path.exists('/proc/device-tree/compatible'):
-        try:
-            with open('/proc/device-tree/compatible', 'rb') as f:
-                content = f.read().split(b"\0")
-                if b"apple,arm-platform" in content:
-                    os.environ["ASAHI_VISIBLE_DEVICES"] = "1"
-                    return "asahi"
-        except OSError:
-            pass
-
-    return None
-
-
-def check_metal(args: ContainerArgType) -> bool:
-    if args.container:
-        return False
-    return platform.system() == "Darwin"
-
-
-@lru_cache(maxsize=1)
-def check_nvidia() -> Optional[Literal["cuda"]]:
-    try:
-        command = ['nvidia-smi', '--query-gpu=index,uuid', '--format=csv,noheader']
-        result = run_cmd(command, encoding="utf-8")
-    except (OSError, subprocess.CalledProcessError):
-        return None
-
-    smi_lines = result.stdout.splitlines()
-    parsed_lines: list[list[str]] = [[item.strip() for item in line.split(',')] for line in smi_lines if line]
-
-    if not parsed_lines:
-        return None
-
-    indices, uuids = map(list, zip(*parsed_lines))
-    # Get the list of devices specified by CUDA_VISIBLE_DEVICES, if any
-    cuda_visible_devices = os.environ.get("CUDA_VISIBLE_DEVICES", "")
-    visible_devices = cuda_visible_devices.split(',') if cuda_visible_devices else []
-    for device in visible_devices:
-        if device not in indices and not any(uuid.startswith(device) for uuid in uuids):
-            perror(f"{device} not found")
-            return None
-
-    configured, unconfigured = find_in_cdi(visible_devices + ["all"])
-
-    configured_has_all = "all" in configured
-    if unconfigured and not configured_has_all:
-        perror(f"No CDI configuration found for {','.join(unconfigured)}")
-        perror("You can use the \"nvidia-ctk cdi generate\" command from the ")
-        perror("nvidia-container-toolkit to generate a CDI configuration.")
-        perror("See ramalama-cuda(7).")
-        return None
-    elif configured:
-        if configured_has_all:
-            configured.remove("all")
-            if not configured:
-                configured = indices
-
-        os.environ["CUDA_VISIBLE_DEVICES"] = ','.join(configured)
-        return "cuda"
-
-    return None
-
-
-def check_ascend() -> Optional[Literal["cann"]]:
-    try:
-        command = ['npu-smi', 'info']
-        run_cmd(command, encoding="utf-8")
-        os.environ["ASCEND_VISIBLE_DEVICES"] = "0"
-        return "cann"
-    except Exception:
-        pass
-
-    return None
-
-
-def check_rocm_amd() -> Optional[Literal["hip"]]:
-    if is_arm():
-        # ROCm is not available for arm64, use Vulkan instead
-        return None
-
-    gpu_num = 0
-    gpu_bytes = 0
-    for i, (np, props) in enumerate(amdkfd.gpus()):
-        # Radeon GPUs older than gfx900 are not supported by ROCm (e.g. Polaris)
-        if props['gfx_target_version'] < 90000:
-            continue
-
-        mem_banks_count = int(props['mem_banks_count'])
-        mem_bytes = 0
-        for bank in range(mem_banks_count):
-            bank_props = amdkfd.parse_props(np + f'/mem_banks/{bank}/properties')
-            # See /usr/include/linux/kfd_sysfs.h for possible heap types
-            #
-            # Count public and private framebuffer memory as VRAM
-            if bank_props['heap_type'] in [amdkfd.HEAP_TYPE_FB_PUBLIC, amdkfd.HEAP_TYPE_FB_PRIVATE]:
-                mem_bytes += int(bank_props['size_in_bytes'])
-
-        if mem_bytes > MIN_VRAM_BYTES and mem_bytes > gpu_bytes:
-            gpu_bytes = mem_bytes
-            gpu_num = i
-
-    if gpu_bytes:
-        os.environ["HIP_VISIBLE_DEVICES"] = str(gpu_num)
-        return "hip"
-
-    return None
-
-
-def is_arm() -> bool:
-    return platform.machine() in ('arm64', 'aarch64')
-
-
-def check_intel() -> Optional[Literal["intel"]]:
-    igpu_num = 0
-
-    if platform.system() == "Windows":
-        igpu_num = _check_intel_windows()
-    else:
-        igpu_num = _check_intel_linux()
-
-    if igpu_num:
-        os.environ["INTEL_VISIBLE_DEVICES"] = str(igpu_num)
-        return "intel"
-
-    return None
-
-
-def _check_intel_linux() -> int:
-    igpu_num = 0
-    # Device IDs for select Intel GPUs.  See: https://dgpu-docs.intel.com/devices/hardware-table.html
-    intel_gpus = (
-        b"0xe20b",
-        b"0xe20c",
-        b"0x46a6",
-        b"0x46a8",
-        b"0x46aa",
-        b"0x56a0",
-        b"0x56a1",
-        b"0x7d51",
-        b"0x7dd5",
-        b"0x7d55",
-    )
-    intel_driver_glob_patterns = ["/sys/bus/pci/drivers/i915/*/device", "/sys/bus/pci/drivers/xe/*/device"]
-    # Check to see if any of the device ids in intel_gpus are in the device id of the i915 / xe driver
-    for fp in sorted([i for p in intel_driver_glob_patterns for i in glob.glob(p)]):
-        with open(fp, 'rb') as file:
-            content = file.read()
-            for gpu_id in intel_gpus:
-                if gpu_id in content:
-                    igpu_num += 1
-    return igpu_num
-
-
-def _check_intel_windows() -> int:
-    """Detect Intel GPUs on Windows via WMI Win32_VideoController."""
-    try:
-        import wmi  # type: ignore
-
-        w = wmi.WMI()
-        return sum(1 for gpu in w.Win32_VideoController() if "intel" in (gpu.Name or "").lower())
-    except Exception:
-        return 0
-
-
-def check_mthreads() -> Optional[Literal["musa"]]:
-    try:
-        command = ['mthreads-gmi']
-        run_cmd(command, encoding="utf-8")
-        os.environ["MUSA_VISIBLE_DEVICES"] = "0"
-        return "musa"
-    except Exception:
-        pass
-
-    return None
-
-
-AccelType: TypeAlias = Literal["asahi", "cuda", "cann", "hip", "intel", "musa"]
-
-
-@lru_cache(maxsize=1)
-def get_accel() -> AccelType | Literal["none"]:
-    checks: tuple[Callable[[], Optional[AccelType]], ...] = (
-        check_asahi,
-        cast(Callable[[], Optional[Literal['cuda']]], check_nvidia),
-        check_ascend,
-        check_rocm_amd,
-        check_intel,
-        check_mthreads,
-    )
-    for check in checks:
-        if result := check():
-            return result
-    return "none"
-
-
-def set_accel_env_vars():
-    if get_accel_env_vars():
-        return
-
-    get_accel()
-
-
-def set_gpu_type_env_vars():
-    if get_gpu_type_env_vars():
-        return
-
-    get_accel()
-
-
-GPUEnvVar: TypeAlias = Literal[
-    "ASAHI_VISIBLE_DEVICES",
-    "ASCEND_VISIBLE_DEVICES",
-    "CUDA_VISIBLE_DEVICES",
-    "GGML_VK_VISIBLE_DEVICES",
-    "HIP_VISIBLE_DEVICES",
-    "INTEL_VISIBLE_DEVICES",
-    "MUSA_VISIBLE_DEVICES",
-]
-
-
-def get_gpu_devices():
-    devices = {}
-    for dev in ["dri", "kfd", "accel"]:
-        path = "/dev/" + dev
-        if os.path.exists(path):
-            devices[dev] = path
-    return dict(sorted(devices.items()))
-
-
-def get_gpu_type_env_vars() -> dict[GPUEnvVar, str]:
-    return {k: v for k in get_args(GPUEnvVar) if (v := os.environ.get(k))}
-
-
-AccelEnvVar: TypeAlias = Literal[
-    "CUDA_LAUNCH_BLOCKING",
-    "HSA_VISIBLE_DEVICES",
-    "HSA_OVERRIDE_GFX_VERSION",
-    "MTHREADS_VISIBLE_DEVICES",
-]
-
-
-def get_accel_env_vars() -> dict[GPUEnvVar | AccelEnvVar, str]:
-    gpu_env_vars: dict[GPUEnvVar, str] = get_gpu_type_env_vars()
-    accel_env_vars: dict[AccelEnvVar, str] = {k: v for k in get_args(AccelEnvVar) if (v := os.environ.get(k))}
-    return gpu_env_vars | accel_env_vars
-
-
 def rm_until_substring(input: str, substring: str) -> str:
     pos = input.find(substring)
     if pos == -1:
@@ -665,57 +326,6 @@ def latest_tagged_image(image: str) -> str:
     if len(image.split(":")) > 1:
         return image
     return f"{image}:latest"
-
-
-class AccelImageArgsWithImage(Protocol):
-    image: str
-
-
-class AccelImageArgsOtherRuntime(Protocol):
-    runtime: str
-    container: bool
-    quiet: bool
-
-
-class AccelImageArgsOtherRuntimeRAG(Protocol):
-    rag: bool
-    runtime: str
-    container: bool
-    quiet: bool
-
-
-AccelImageArgs: TypeAlias = Union[None, AccelImageArgsOtherRuntime, AccelImageArgsOtherRuntimeRAG]
-
-
-def accel_image(config: Config, images: Optional[dict[str, str]] = None, conf_key: str = "image") -> str:
-    """
-    Selects the appropriate image based on config, arguments, environment.
-    "images" is a mapping of environment variable names to image names. If not specified,
-    the runtime plugin is asked to select the image.
-    "conf_key" is the configuration key that holds the configured value of the selected image.
-    If not specified, it defaults to "image".
-
-    Never pulls images; callers that need pulling should pass the result to ensure_image().
-    """
-
-    # User provided an image via config; tag with :latest if no tag given
-    if config.is_set(conf_key):
-        return latest_tagged_image(getattr(config, conf_key))
-
-    set_gpu_type_env_vars()
-    gpu_type = next(iter(get_gpu_type_env_vars()), "")
-
-    if not images:
-        # Ask the runtime plugin to select the image based on detected GPU and its own logic
-        from ramalama.plugins.loader import get_runtime
-
-        plugin_image = get_runtime(config.runtime).get_container_image(config, gpu_type)
-        if plugin_image is not None:
-            return latest_tagged_image(plugin_image)
-        images = config.images  # plugin returned None (e.g., MLX); fall back to user dict
-
-    # Explicit images dict provided (e.g., RAG): select by detected GPU type
-    return latest_tagged_image(images.get(gpu_type, getattr(config, f"default_{conf_key}")))
 
 
 def ensure_image(conman: Optional[str], image: str, should_pull: bool = False) -> str:
